commontest/requirements.txt
requests>=2.25.1
flask~=2.1.0
itsdangerous==1.1.0
flask-session==0.3.2
flask-login==0.5.0
flask-wtf==0.14.3
apscheduler==3.8.1
msal==1.10.0
cassandra-driver==3.20.2
msrestazure==0.6.4
azure-keyvault-secrets==4.2.0
azure-identity==1.5.0
azure-graphrbac==0.60.0
pytz==2021.3
python-dateutil==2.8.2
msgraph-core


Werkzeug~=2.0.0
pydantic[email]~=1.9.1

pandas
openpyxl